restructure contracts package
