+++ packages/contracts/src/contracts/core/mod.rs
@@ -1,0 +1,1 @@
+pub mod verifier;
