+++ packages/contracts/src/contracts/mod.rs
@@ -1,0 +1,6 @@
+pub mod core;
+
+#[cfg(any(
+    feature = "precompile-test-contract",
+))]
+pub mod test_contracts;
