+++ packages/contracts/src/contracts/test_contracts/mod.rs
@@ -1,0 +1,1 @@
+pub mod precompile;
