--- packages/contracts/src/mocks/mod.rs
@@ -1,3 +1,0 @@
-#![cfg(feature = "precompile-test-contract")]
-
-pub mod backends;
